requirements.txt
pandas
numpy
pillow
scikit-image
scikit-learn
pickle5
